perf(python): defer runtime type definitions
