+++ py/bindings/python/nanocodex/_types.py
@@ -1,0 +1,21 @@
+from typing import TypedDict
+
+
+class EstimatedCost(TypedDict):
+    usd: str
+    input_usd: str
+    cached_input_usd: str
+    cache_write_input_usd: str
+    output_usd: str
+    service_tier: str
+
+
+class Usage(TypedDict):
+    input_tokens: int
+    cached_input_tokens: int
+    cache_write_input_tokens: int
+    output_tokens: int
+    reasoning_output_tokens: int
+    total_tokens: int
+    estimated_cost: EstimatedCost | None
+    cost_status: str
